# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dochub API
  version: "1.0.0"
  description: |
    Organizational RAG document hub.

    **Nothing reaches the backend while a user is choosing sources.** The browser holds
    repository details, folder ids and local files until Process is pressed.

    **Process contract.** `POST /artifacts/{artifactId}/process` is the only call the
    Upload screen makes. For each source it writes a `source_documents` row with status
    `RequestUpload` and puts the source on the extract queue
    (`dochub-source-upload-requested`). The extractor service claims one message at a
    time, flips the row to `Uploading`, moves the files to Azure Blob storage under
    `teams/{team}/groups/{group}/artifacts/{artifact}/{uploadedAt}`, sets `Uploaded`, and
    hands the documents to the process queue (`dochub-document-process-requested`). For
    every blob it stores the content MD5 Azure returned in a `document_versions` row.
    The RAG platform (`ragplatform/`) consumes the process queue, indexes each document
    and reports back through the **Ingestion** endpoints: once per document, then once
    for the artifact, which closes the source and sets the artifact's status.

    Local files are the one exception that needs a prior call: their bytes go to
    `POST /uploads/staging` immediately before Process, and the returned staging ids are
    passed in the source's `options`.

    **Recurring updates.** A SharePoint or Google Drive source may carry a `schedule`,
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /api/auth/sso:
    post:
      tags: [Auth]
      summary: Exchange a Google or Microsoft ID token for a Dochub access token
      description: New users are created on first sign-in but hold no organization until an owner adds them.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, idToken]
              properties:
                provider:
                  type: string
                  enum: [google, microsoft, dev]
                idToken:
                  type: string
                  description: The provider's ID token. For `dev` (local only), `email|Display Name`.
                organizationId: { type: string, format: uuid, nullable: true }
      responses:
        "200": { description: Signed in, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "401": { $ref: "#/components/responses/Error" }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user, active organization and memberships
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "401": { $ref: "#/components/responses/Error" }
  /api/auth/switch-organization:
    post:
      tags: [Auth]
      summary: Re-issue the access token scoped to another organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId]
              properties:
                organizationId: { type: string, format: uuid }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "403": { $ref: "#/components/responses/Error" }
...
  /api/chat/conversations:
    get:
      tags: [Chat]
      summary: The caller's conversations in the active organization, newest first
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ChatConversation" } }
    post:
      tags: [Chat]
      summary: Start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, nullable: true }
                scope: { $ref: "#/components/schemas/ChatScope" }
                scopeId: { type: string, format: uuid, nullable: true, description: Required unless scope is Organization. }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/ChatConversation" } } }
        "400": { $ref: "#/components/responses/Error" }
        "404":
          description: "`scope_not_found` — no such team, group or artifact in this organization."
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /api/chat/conversations/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Chat]
      summary: A conversation with its messages and their sources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation: { $ref: "#/components/schemas/ChatConversation" }
                  messages: { type: array, items: { $ref: "#/components/schemas/ChatMessage" } }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [Chat]
      summary: Rename a conversation or change what it searches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, nullable: true }
                scope: { $ref: "#/components/schemas/ChatScope" }
                scopeId: { type: string, format: uuid, nullable: true }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/ChatConversation" } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [Chat]
      summary: Delete a conversation and its messages
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
  /api/chat/conversations/{id}/messages:
    post:
      tags: [Chat]
      summary: Ask a question
      description: |
        Saves the question, then streams `text/event-stream`. Each event is one
        `data:` line holding a JSON object with a `type`:

        - `started` — `{questionId, answerId, scope, title}`
        - `sources` — `{query, sources: ChatSource[]}`, before any text
        - `delta` — `{text}`, answer text as it is written
        - `done` — `{cited: int[], model, usage}`
        - `error` — `{error, answerId}` when the knowledge base fails mid-answer

        The answer is saved with its sources even if the stream fails or the caller leaves.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1, maxLength: 8000 }
      responses:
        "200":
          description: The answer stream.
          content:
            text/event-stream:
              schema: { type: string, example: "data: {\"type\":\"delta\",\"text\":\"Claims are acknowledged \"}\n\n" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from `POST /api/auth/sso`. Carries the active `org_id`, so every request is tenant-scoped by the token.
    serviceKey:
      type: apiKey
      in: header
      name: X-Dochub-Service-Key
      description: Equal to `Ingestion:ServiceKey` (≥32 chars). Used only by the RAG platform.
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: owner_only }
        message: { type: string }
        details: { nullable: true }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
        identityProvider: { type: string, example: google }
        canCreateOrganizations: { type: boolean }
        role: { type: string, enum: [Member, Admin, Owner], nullable: true }
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        initials: { type: string }
        plan: { type: string }
        role: { type: string, enum: [Member, Admin, Owner] }
        memberCount: { type: integer }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        expiresAt: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
        organization: { allOf: [{ $ref: "#/components/schemas/Organization" }], nullable: true }
        organizations: { type: array, items: { $ref: "#/components/schemas/Organization" } }
...
    ChatScope:
      type: string
      enum: [Organization, Team, Group, Artifact]
    ChatConversation:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        scope: { $ref: "#/components/schemas/ChatScope" }
        scopeId: { type: string, format: uuid, nullable: true }
        scopeLabel: { type: string, example: "Claims › Claims IT › Claims Core Repo" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ChatSource:
      type: object
      description: One numbered passage an answer was written from, as the RAG platform returned it.
      properties:
        n: { type: integer, example: 1 }
        chunk_id: { type: string }
        document_id: { type: string }
        dochub_document_id: { type: string, format: uuid, nullable: true }
        artifact_id: { type: string, format: uuid, nullable: true }
        title: { type: string, nullable: true }
        filename: { type: string, nullable: true }
        heading_path: { type: array, items: { type: string } }
        location: { type: string, nullable: true, example: "page 3" }
        chunk_type: { type: string, example: prose }
        snippet: { type: string }
        score: { type: number }
    ChatMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [User, Assistant] }
        content: { type: string }
        sources: { type: array, nullable: true, items: { $ref: "#/components/schemas/ChatSource" } }
        cited: { type: array, nullable: true, items: { type: integer } }
        model: { type: string, nullable: true }
        error: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...